# shared/docs/swagger.json
{"openapi":"3.0.0","paths":{"/tasks":{"get":{"operationId":"TaskController_get","parameters":[{"name":"projectId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetTaskUseCaseResponseDTO"}}}}}},"tags":["Task (Tarefas)"]},"post":{"operationId":"TaskController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskUseCaseRequestDTO"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskUseCaseResponseDTO"}}}}},"tags":["Task (Tarefas)"]}},"/tasks/{id}":{"patch":{"operationId":"TaskController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskUseCaseRequestDTO"}}}},"responses":{"204":{"description":""}},"tags":["Task (Tarefas)"]},"get":{"operationId":"TaskController_getDetail","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetDetailTaskUseCaseResponseDTO"}}}}},"tags":["Task (Tarefas)"]}},"/projects":{"get":{"operationId":"ProjectController_get","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetProjectUseCaseResponseDTO"}}}}}},"tags":["Project (Projeto)"]},"post":{"operationId":"ProjectController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectUseCaseRequestDTO"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectUseCaseResponseDTO"}}}}},"tags":["Project (Projeto)"]}},"/projects/{id}":{"patch":{"operationId":"ProjectController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectUseCaseRequestDTO"}}}},"responses":{"204":{"description":""}},"tags":["Project (Projeto)"]},"get":{"operationId":"ProjectController_getDetail","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetDetailProjectUseCaseResponseDTO"}}}}},"tags":["Project (Projeto)"]}}},"info":{"title":"Artis - API","description":"Artis - API","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://127.0.0.1:3000","description":"URL Local"}],"components":{"schemas":{"Status":{"type":"object","properties":{"id":{"type":"string"},"description":{"type":"string"},"type":{"type":"string"}},"required":["id","description","type"]},"GetTaskUseCaseResponseDTO":{"type":"object","properties":{"id":{"type":"string"},"text":{"type":"string"},"title":{"type":"string"},"startDate":{"format":"date-time","type":"string"},"endDate":{"format":"date-time","type":"string"},"statusId":{"type":"string"},"Status":{"$ref":"#/components/schemas/Status"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","text","title","startDate","endDate","statusId","Status","createdAt","updatedAt"]},"CreateTaskUseCaseRequestDTO":{"type":"object","properties":{"text":{"type":"string"},"title":{"type":"string"},"endDate":{"format":"date-time","type":"string"},"statusId":{"type":"string"},"projectId":{"type":"string"}},"required":["text","title","endDate","statusId","projectId"]},"CreateTaskUseCaseResponseDTO":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]},"UpdateTaskUseCaseRequestDTO":{"type":"object","properties":{"text":{"type":"string"},"title":{"type":"string"},"endDate":{"format":"date-time","type":"string"},"statusId":{"type":"string"},"Status":{"$ref":"#/components/schemas/Status"}},"required":["text","title","endDate","statusId","Status"]},"GetDetailTaskUseCaseResponseDTO":{"type":"object","properties":{"id":{"type":"string"},"text":{"type":"string"},"title":{"type":"string"},"startDate":{"format":"date-time","type":"string"},"endDate":{"format":"date-time","type":"string"},"statusId":{"type":"string"},"Status":{"$ref":"#/components/schemas/Status"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","text","title","startDate","endDate","statusId","Status","createdAt","updatedAt"]},"GetProjectUseCaseResponseDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"startDate":{"format":"date-time","type":"string"},"endDate":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"completionPercentage":{"type":"number"},"isDelayed":{"type":"boolean"}},"required":["id","name","description","startDate","endDate","createdAt","updatedAt","completionPercentage","isDelayed"]},"CreateProjectUseCaseRequestDTO":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"endDate":{"format":"date-time","type":"string"}},"required":["name","description","endDate"]},"CreateProjectUseCaseResponseDTO":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]},"UpdateProjectUseCaseRequestDTO":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"endDate":{"format":"date-time","type":"string"}},"required":["name","description","endDate"]},"GetDetailProjectUseCaseResponseDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"startDate":{"format":"date-time","type":"string"},"endDate":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"completionPercentage":{"type":"number"},"isDelayed":{"type":"boolean"}},"required":["id","name","description","startDate","endDate","createdAt","updatedAt","completionPercentage","isDelayed"]}}}}