# src/modules/task/use-cases/create-task/create-task.use-case.schema.ts
import { z } from 'zod';

// Define o schema para a requisição
export const CreateTaskUseCaseRequestDTOSchema = z.object({
  title: z.string().min(1, 'Name is required'),
  projectId: z.string().min(1, 'ProjectId is required'),
  text: z.string().optional(),
  endDate: z
    .string()
    .refine((val) => !isNaN(Date.parse(val)), 'Invalid end date'),
});